Split state between the uploader interface and the imports information
