+++ client/app/containers/ImportDeckFormContainer.jsx
@@ -1,12 +1,12 @@
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import ImportDeckForm from 'presentationals/components/deckManagement/ImportDeckForm';
-import { importUploadError } from 'actions/userImportsActions';
+import { importUploadError } from 'actions/fineUploaderActions';
 
 function mapStateToProps(state) {
   return {
     authState: state.app.authentication,
-    userImportState: state.app.userImports,
+    fineUploaderState: state.app.fineUploader,
   };
 }
 function mapDispatchToProps(dispatch) {
