+++ client/app/reducers/app/index.js
@@ -7,6 +7,7 @@ import feed from 'reducers/app/feedReducer';
 import annotations from 'reducers/app/annotations';
 import deckManagement from 'reducers/app/deckManagementReducer';
 import userImports from 'reducers/app/userImportsReducer';
+import fineUploader from 'reducers/app/fineUploaderReducer';
 import deckImport from 'reducers/app/deckImportReducer';
 import presentationView from 'reducers/app/presentationViewReducer';
 
@@ -18,6 +19,7 @@ const app = combineReducers({
   annotations,
   deckManagement,
   userImports,
+  fineUploader,
   deckImport,
   presentationView,
 });
